% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjo_palettes.R
\docType{data}
\name{pjo_palettes}
\alias{pjo_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
pjo_palettes
}
\description{
For the time being, all palettes only contain 5 colours.
Palettes are derived using coolor.co and making personal adjustments.
}
\keyword{datasets}
