% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjo_palettes.R
\name{scale_color_pjo}
\alias{scale_color_pjo}
\title{PJO colour scale}
\usage{
scale_color_pjo(palette = "LightningThief", discrete = TRUE, ...)
}
\arguments{
\item{palette}{Which palette in pjo_palettes to use for colour scale (default = 'LightningThief')}

\item{discrete}{logical. TRUE if the colour aesthetic is discrete or FALSE if the colour aesthetic is continuous}

\item{...}{Additional arguments}
}
\description{
PJO colour scale
}
