% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjo_palettes.R
\name{scale_fill_pjo}
\alias{scale_fill_pjo}
\title{PJO fill scale}
\usage{
scale_fill_pjo(palette = "LightningThief", discrete = TRUE, ...)
}
\arguments{
\item{palette}{Which palette in \code{pjo_palettes} to use for fill scale (default = 'LightningThief')}

\item{discrete}{logical. TRUE if the fill aesthetic is discrete or FALSE if the fill aesthetic is continuous}

\item{...}{Additional arguments}
}
\description{
PJO fill scale
}
