% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_pjo_palettes.R
\name{view_pjo_palette}
\alias{view_pjo_palette}
\title{A PJO palette generator}
\usage{
view_pjo_palette(name, n)
}
\arguments{
\item{name}{Name of desired palette.}

\item{n}{Number of colors desired.}
}
\value{
A vector of colours.
}
\description{
A PJO palette generator
}
\keyword{colors}
